+++ src/main/java/com/sleroux/bank/dao/IOperationDao.java
@@ -11,18 +11,18 @@ public interface IOperationDao extends IOperations<Operation> {
 
 	int insertIgnore(Operation _o);
 
-	List<Operation> findUncategorized(int _userID);
+	List<Operation> findUncategorized(long _userID);
 
-	List<String> getCategoriesDebit(int _userID);
+	List<String> getCategoriesDebit(long _userID);
 
-	List<String> getCategoriesCredit(int _userID);
+	List<String> getCategoriesCredit(long _userID);
 
 	List<String> getSuggestionsFor(Operation _o);
 
-	List<AccountBalance> getSoldes(int _userID);
+	List<AccountBalance> getSoldes(long _userID);
 
-	List<AggregatedOperations> findAggregatedYearMonth(int _year, int _month, int _userID);
+	List<AggregatedOperations> findAggregatedYearMonth(int _year, int _month, long _userID);
 
-	List<Operation> findByCategoYearMonth(Integer year, Integer month, String catego, int _userID);
+	List<Operation> findByCategoYearMonth(Integer year, Integer month, String catego, long _userID);
 
 }
