+++ src/main/java/com/sleroux/bank/dao/impl/BudgetDao.java
@@ -24,7 +24,7 @@ public class BudgetDao extends AbstractHibernateDao<Budget> implements IBudgetDa
 	}
 
 	@Override
-	public List<String> getCredits(int _userID) {
+	public List<String> getCredits(long _userID) {
 		Query q = getCurrentSession().createQuery(
 				"select distinct catego from Budget where credit > 0 and compte = :compte and user_id = :user_id order by catego");
 		q.setParameter("user_id", _userID);
@@ -33,7 +33,7 @@ public class BudgetDao extends AbstractHibernateDao<Budget> implements IBudgetDa
 	}
 
 	@Override
-	public List<String> getDebits(int _userID) {
+	public List<String> getDebits(long _userID) {
 		Query q = getCurrentSession().createQuery(
 				"select distinct catego from Budget where debit > 0 and compte = :compte and user_id = :user_id order by catego");
 		q.setParameter("user_id", _userID);
@@ -42,7 +42,7 @@ public class BudgetDao extends AbstractHibernateDao<Budget> implements IBudgetDa
 	}
 
 	@Override
-	public List<Integer> getYears(int _userID) {
+	public List<Integer> getYears(long _userID) {
 		String sql = "select distinct year from Budget where user_id = :user_id order by year";
 		Query q = getCurrentSession().createQuery(sql);
 		q.setParameter("user_id", _userID);
@@ -50,13 +50,13 @@ public class BudgetDao extends AbstractHibernateDao<Budget> implements IBudgetDa
 	}
 
 	@Override
-	public List<String> getComptesEpargne(int _userID) {
+	public List<String> getComptesEpargne(long _userID) {
 		String sql = "select distinct compte from Budget where compte <> 'COURANT' and user_id = :user_id order by compte";
 		return getCurrentSession().createQuery(sql).setParameter("user_id", _userID).list();
 	}
 
 	@Override
-	public List<Budget> getMonthDebits(Integer _year, int _month, int _userID) {
+	public List<Budget> getMonthDebits(Integer _year, int _month, long _userID) {
 		Query q = getCurrentSession().createSQLQuery("call bank.getMonthDebits(:year,:month,:user_id)");
 		q.setParameter("year", _year);
 		q.setParameter("month", _month);
@@ -67,7 +67,7 @@ public class BudgetDao extends AbstractHibernateDao<Budget> implements IBudgetDa
 	}
 
 	@Override
-	public List<Budget> getMonthCredits(Integer _year, int _month, int _userID) {
+	public List<Budget> getMonthCredits(Integer _year, int _month, long _userID) {
 		Query q = getCurrentSession().createSQLQuery("call bank.getMonthCredits(:year,:month,:user_id)");
 		q.setParameter("year", _year);
 		q.setParameter("month", _month);
@@ -77,7 +77,7 @@ public class BudgetDao extends AbstractHibernateDao<Budget> implements IBudgetDa
 	}
 
 	@Override
-	public Budget getBudgetForCompte(String _compte, Integer _year, int _month, int _userID) {
+	public Budget getBudgetForCompte(String _compte, Integer _year, int _month, long _userID) {
 		Query q = getCurrentSession().createSQLQuery(
 				"select sum(debit) debit,sum(credit) credit from budget where compte = :compte and year= :year and month = :month and user_id = :user_id");
 		q.setParameter("compte", _compte);
@@ -100,7 +100,7 @@ public class BudgetDao extends AbstractHibernateDao<Budget> implements IBudgetDa
 	}
 
 	@Override
-	public List<Changes> getAdded(int _userID) {
+	public List<Changes> getAdded(long _userID) {
 		String sql = "select a.id as id, a.year as year, a.month as month, a.catego as catego, a.debit as debit, a.credit as credit, a.compte as compte, a.notes as notes from budget a left join (select * from budget_backup inner join (select max(timestamp) last_backup from budget_backup) t on t.last_backup = timestamp) b on a.year = b.year and a.month = b.month and a.catego = b.catego and a.compte = b.compte where  b.timestamp is null";
 		Query q = getCurrentSession().createSQLQuery(sql);
 		q.setResultTransformer(Transformers.aliasToBean(Changes.class));
@@ -108,7 +108,7 @@ public class BudgetDao extends AbstractHibernateDao<Budget> implements IBudgetDa
 	}
 
 	@Override
-	public List<Changes> getUpdated(int _userID) {
+	public List<Changes> getUpdated(long _userID) {
 		String sql = "select a.year year, a.month month, a.catego catego, a.compte compte, a.debit debit, a.credit credit, a.notes, b.debit oldDebit, b.credit oldCredit from budget a inner join budget_backup b on a.year = b.year and a.month = b.month and a.catego = b.catego and a.compte = b.compte inner join (select max(timestamp) last_backup from budget_backup) t on b.timestamp = t.last_backup where a.debit <> b.debit or a.credit <> b.credit;";
 		Query q = getCurrentSession().createSQLQuery(sql);
 		q.setResultTransformer(Transformers.aliasToBean(Changes.class));
@@ -116,7 +116,7 @@ public class BudgetDao extends AbstractHibernateDao<Budget> implements IBudgetDa
 	}
 
 	@Override
-	public List<Changes> getDeleted(int _userID) {
+	public List<Changes> getDeleted(long _userID) {
 		String sql = "select a.year, a.month, a.catego, a.compte, a.debit, a.credit, a.notes from budget_backup a inner join (select max(timestamp) last_backup from budget_backup) t on a.timestamp = t.last_backup left join budget b on a.year = b.year and a.month = b.month and a.catego = b.catego and a.compte = b.compte where  b.id is null";
 		Query q = getCurrentSession().createSQLQuery(sql);
 		q.setResultTransformer(Transformers.aliasToBean(Changes.class));
@@ -124,7 +124,7 @@ public class BudgetDao extends AbstractHibernateDao<Budget> implements IBudgetDa
 	}
 
 	@Override
-	public void saveChange(Changes _change, int _userID) {
+	public void saveChange(Changes _change, long _userID) {
 
 		StringBuilder sql = new StringBuilder();
 		sql.append(
@@ -155,7 +155,7 @@ public class BudgetDao extends AbstractHibernateDao<Budget> implements IBudgetDa
 	}
 
 	@Override
-	public Budget findByYearMonthCatego(int _year, int _month, String _catego, String _compte, int _userID) {
+	public Budget findByYearMonthCatego(int _year, int _month, String _catego, String _compte, long _userID) {
 		return (Budget) getCurrentSession()
 				.createQuery(
 						"from Budget where year = :year and month = :month and catego = :catego and compte=:compte and user_id=:user_id")
@@ -165,7 +165,7 @@ public class BudgetDao extends AbstractHibernateDao<Budget> implements IBudgetDa
 	}
 
 	@Override
-	public List<AggregatedOperations> findBudgetForMonth(int _year, int _month, int _userID) {
+	public List<AggregatedOperations> findBudgetForMonth(int _year, int _month, long _userID) {
 		Query query = getCurrentSession().createSQLQuery(
 				"select year, month, compte as account, catego, sum(credit) as credit, sum(debit) as debit from budget where year = :year and month = :month and user_id = :user_id group by year, month, compte, catego");
 		query.setParameter("year", _year);
@@ -176,7 +176,7 @@ public class BudgetDao extends AbstractHibernateDao<Budget> implements IBudgetDa
 	}
 
 	@Override
-	public BigDecimal getEstimatedEndOfMonthBalance(int _year, int _month, int _userID) {
+	public BigDecimal getEstimatedEndOfMonthBalance(int _year, int _month, long _userID) {
 		Query query = getCurrentSession().createSQLQuery(
 				"select sum(credit) - sum(debit)  from budget where user_id = :user_id and compte = 'COURANT' and date(concat(year,'-', month,'-',1)) <=  date(concat(:year,'-',:month,'-',1))");
 		query.setParameter("year", _year);
